Refactor/polish code, UI and improve performance
- swap out text output field for text label
  - workaround observed performance issues with using the
    widget.Entry type for output text by using the widget.Label
    type and relying solely on the Copy to Clipboard button to
    retrieve decoded input text
- refactor/cleanup code (doc comments, func names, organization)
- move repeated text strings to constants
- perform explicit garbage collection attempt when the Reset
  button is used
- remove window resize constraint (allow user resizable window)
